// src/app/api/models/ticket-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserSummary } from '../models/user-summary';
export interface TicketResponse {
  aiSummary?: string;
  assignedTo?: UserSummary;
  category?: string;
  createdAt?: string;
  createdBy?: UserSummary;
  description?: string;
  id?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  status?: 'OPEN' | 'IN_PROGRESS' | 'RESOLVED' | 'CLOSED';
  title?: string;
  updatedAt?: string;
}